Moto Genie/requirements.txt
Flask==2.2.2
pandas==1.5.3
requests==2.28.2
google-generativeai==0.2.0
fpdf==1.7.2
gunicorn==20.1.0

